requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
fonttools==4.28.3
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.6.4
matplotlib==3.5.0
multitasking==0.0.10
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
threadpoolctl==3.0.0
tomli==1.2.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wincertstore==0.2
wrapt==1.13.3
yfinance==0.1.66
